vscode: refactor
