+++ tools/vscode-extension/src/coverageHelper.ts
@@ -18,7 +18,8 @@ import * as vscode from 'vscode';
 import {Uri} from 'vscode';
 import {println} from './logger';
 import {getApi, FileDownloader} from '@microsoft/vscode-file-downloader-api';
-import {extensionConfig} from './config';
+import {extensionConfig} from './config'
+import {getOSSFuzzCloudURL, getLocalOutBuildDir, downloadRemoteURL} from './utils';
 
 const path = require('path');
 let isCodeCoverageEnabled = false;
@@ -69,44 +70,13 @@ export async function compareLocalToRemoteCoverage(
   println("Checking the file matching");
   /* Get the coverage from the remote server */
   const fileDownloader: FileDownloader = await getApi();
-
-  const currentDate = new Date();
-  const yesterday = new Date(currentDate);
-  yesterday.setDate(yesterday.getDate() - 1);
-
-  const day = yesterday.getDate();
-  const month = yesterday.getMonth();
-  const year = yesterday.getFullYear();
-
-  var urlString = 'https://storage.googleapis.com/oss-fuzz-coverage/' +
-  projectName +
-  '/reports/' +
-  year.toString();
-
-  if (month < 10) {
-    urlString += "0";
-  }
-  urlString += month.toString();
-  if (day < 10) {
-    urlString += "0";
-  }
-  urlString += day.toString();
-  urlString += '/linux/summary.json';
+  var urlString = await getOSSFuzzCloudURL(projectName) + '/linux/summary.json';
 
   println("URL: " + urlString);
-  var codeCoverageFile: vscode.Uri;
-  try {
-    codeCoverageFile = await fileDownloader.downloadFile(
-      vscode.Uri.parse(
-        urlString
-      ),
-      'summary.json',
-      context
-    );
-
-  } catch (err) {
+  var codeCoverageFile: false | vscode.Uri = await downloadRemoteURL(urlString, 'summary.json', context);
+  if (!codeCoverageFile) {
     println(
-      'Could not get the URL. Currently, this feature is only supported for Python projects'
+      'Could not get the coverage summary file'
     );
     return;
   }
@@ -114,22 +84,23 @@ export async function compareLocalToRemoteCoverage(
 
   /* Get the local coverage report */
   // Compare the local coverage to the upstream coverage
-  const summaryCovPath =
-    extensionConfig.ossFuzzPepositoryWorkPath +
-    '/build/out/' +
-    projectName +
-    '/report/linux/summary.json';
-  const localCodeCoverage = await loadSummaryJsonCoverage(context, vscode.Uri.file(summaryCovPath));
+  const localSummaryCovPath = await getLocalOutBuildDir(projectName) + '/report/linux/summary.json';
+  const localCodeCoverage = await loadSummaryJsonCoverage(context, vscode.Uri.file(localSummaryCovPath));
 
   for (var i = 0; i < localCodeCoverage.data[0].files.length; i++) {
     for (var j = 0; j < remoteCoverage.data[0].files.length; j++) {
-      if (localCodeCoverage.data[0].files[i].filename == remoteCoverage.data[0].files[j].filename) {
-        const remoteFuncCount = remoteCoverage.data[0].files[j].summary.functions.count;
-        const localFuncCount = localCodeCoverage.data[0].files[i].summary.functions.count;
+      // Get the file dictionary
+      const localFileData = localCodeCoverage.data[0].files[i];
+      const remoteFileData = remoteCoverage.data[0].files[j];
+
+      // If the filepaths are the same, then we match coverage data
+      if (localFileData.filename == remoteFileData.filename) {
+        const remoteFuncCount = remoteFileData.summary.functions.count;
+        const localFuncCount = localFileData.summary.functions.count;
 
         if (localFuncCount > remoteFuncCount) {
           println("Coverage improved in :" +
-            localCodeCoverage.data[0].files[i].filename +
+          localFileData.filename +
             " [" + localFuncCount + " : " + remoteFuncCount + "]");
         }
       }
