+++ tools/vscode-extension/src/utils.ts
@@ -15,10 +15,71 @@
 ////////////////////////////////////////////////////////////////////////////////
 
 import * as vscode from 'vscode';
+import {extensionConfig} from './config'
+import {getApi, FileDownloader} from '@microsoft/vscode-file-downloader-api';
+
 const fs = require('fs');
 const {spawn} = require('node:child_process');
 
 import {println, printRaw, debugPrintln} from './logger';
+
+export async function downloadRemoteURL(urlString: string, targetFile: string, context: vscode.ExtensionContext,) {
+  const fileDownloader: FileDownloader = await getApi();
+  //var urlString = await getOSSFuzzCloudURL(projectName) + '/linux/summary.json';
+
+  println("URL: " + urlString);
+  var codeCoverageFile: vscode.Uri;
+  try {
+    codeCoverageFile = await fileDownloader.downloadFile(
+      vscode.Uri.parse(
+        urlString
+      ),
+      targetFile,
+      context
+    );
+
+  } catch (err) {
+    println(
+      'Could not get the coverage summary file'
+    );
+    return false;
+  }
+  return codeCoverageFile;
+}
+
+export async function getLocalOutBuildDir(projectName: string) {
+  const summaryCovPath =
+    extensionConfig.ossFuzzPepositoryWorkPath +
+    '/build/out/' +
+    projectName;
+    return summaryCovPath;
+}
+
+export async function getOSSFuzzCloudURL(projectName: string) {
+  const currentDate = new Date();
+  const yesterday = new Date(currentDate);
+  yesterday.setDate(yesterday.getDate() - 1);
+
+  const day = yesterday.getDate();
+  const month = yesterday.getMonth();
+  const year = yesterday.getFullYear();
+
+  var urlString = 'https://storage.googleapis.com/oss-fuzz-coverage/' +
+  projectName +
+  '/reports/' +
+  year.toString();
+
+  if (month < 10) {
+    urlString += "0";
+  }
+  urlString += month.toString();
+  if (day < 10) {
+    urlString += "0";
+  }
+  urlString += day.toString();
+
+  return urlString;
+}
 
 /**
  * Checks if the current workspace has a generated OSS-Fuzz folder. This is the
